move all the jsons to data folder
