+++ pages/events.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import EventCard from '@/components/EventCard';
 import styles from '@/styles/EventCard.module.css';
-import events from '@/pages/Events.json';
+import events from '@/data/Events.json';
 
 export default function EventsPage () {
 	return (
